editing: extract TileEditService from ObjectCommandController
Continues the ObjectCommandController split: floor/roof tile editing
(applyTileChanges, registerTileEdit, applyTileEdit) moves into a focused
TileEditService built on the shared UndoBatcher. ObjectCommandController keeps the
public methods as thin delegations, so callers and the test suite are unchanged.
The tile-only callbacks (ensureElevationTiles, getCurrentElevation,
updateTileSprite) move into the service, and the now-unused TileUtils/unordered_map
includes are dropped from the controller. 235 tests pass.
