+++ src/CMakeLists.txt
@@ -472,6 +472,8 @@ add_library(gecko_editing STATIC
     ui/editing/ObjectCommandController.cpp
     ui/editing/UndoBatcher.h
     ui/editing/UndoBatcher.cpp
+    ui/editing/TileEditService.h
+    ui/editing/TileEditService.cpp
     scripting/MapScriptApi.h
     scripting/MapScriptApi.cpp
     scripting/ScriptApiReference.h
