+++ src/ui/editing/ObjectCommandController.cpp
@@ -7,12 +7,10 @@
 #include "format/map/Tile.h"
 #include "resource/GameResources.h"
 #include "editor/TileChange.h"
-#include "util/TileUtils.h"
 #include "util/UndoStack.h"
 #include "ui/rendering/MapSpriteLoader.h"
 
 #include <algorithm>
-#include <unordered_map>
 #include <spdlog/spdlog.h>
 
 namespace geck {
@@ -37,73 +35,21 @@ ObjectCommandController::ObjectCommandController(resource::GameResources& resour
     , _objects(objects)
     , _wallBlockerOverlays(wallBlockerOverlays)
     , _batcher(undoStack, std::move(onStackChanged))
+    , _tileService(map, _batcher, std::move(ensureElevationTiles), std::move(getCurrentElevation), std::move(updateTileSprite))
     , _refreshObjects(std::move(refreshObjects))
-    , _ensureElevationTiles(std::move(ensureElevationTiles))
-    , _getCurrentElevation(std::move(getCurrentElevation))
-    , _updateTileSprite(std::move(updateTileSprite))
     , _reloadTiles(std::move(reloadTiles)) {
 }
 
 void ObjectCommandController::applyTileChanges(const std::vector<TileChange>& changes, bool applyAfterState) {
-    if (!_map) {
-        return;
-    }
-
-    std::unordered_map<int, std::vector<const TileChange*>> changesByElevation;
-    for (const auto& change : changes) {
-        changesByElevation[change.elevation].push_back(&change);
-    }
-
-    // Apply all changes first, then update sprites in batch.
-    for (const auto& [elevation, elevChanges] : changesByElevation) {
-        auto& elevationTiles = _ensureElevationTiles(elevation);
-
-        for (const auto* change : elevChanges) {
-            if (change->tileIndex < 0 || change->tileIndex >= static_cast<int>(elevationTiles.size())) {
-                continue;
-            }
-
-            uint16_t value = applyAfterState ? change->after : change->before;
-            if (change->isRoof) {
-                elevationTiles[change->tileIndex].setRoof(value);
-            } else {
-                elevationTiles[change->tileIndex].setFloor(value);
-            }
-        }
-
-        if (elevation == _getCurrentElevation()) {
-            for (const auto* change : elevChanges) {
-                if (change->tileIndex >= 0 && change->tileIndex < static_cast<int>(elevationTiles.size())) {
-                    int hexIndex = tileIndexToHexIndex(change->tileIndex);
-                    _updateTileSprite(hexIndex, change->isRoof, elevation);
-                }
-            }
-        }
-    }
+    _tileService.applyTileChanges(changes, applyAfterState);
 }
 
 void ObjectCommandController::registerTileEdit(const std::string& description, const std::vector<TileChange>& changes) {
-    if (changes.empty()) {
-        return;
-    }
-
-    UndoCommand cmd;
-    cmd.description = description;
-    cmd.undo = [this, changes]() {
-        applyTileChanges(changes, false);
-    };
-    cmd.redo = [this, changes]() {
-        applyTileChanges(changes, true);
-    };
-    pushCommand(std::move(cmd));
+    _tileService.registerTileEdit(description, changes);
 }
 
 void ObjectCommandController::applyTileEdit(const std::string& description, const std::vector<TileChange>& changes) {
-    if (changes.empty()) {
-        return;
-    }
-    applyTileChanges(changes, true); // apply the move now and refresh the affected tile sprites
-    registerTileEdit(description, changes);
+    _tileService.applyTileEdit(description, changes);
 }
 
 void ObjectCommandController::addObjectData(const std::shared_ptr<MapObject>& mapObject) {
